doc/pages: changed @addtogroup to @ingroup.¹
1: This ensures group structure is built. @addtogroup wasn't
*defining* the group, apparently, resulting in no structure
for groups that had no explicit body for the group (`@{` & `@}`).
The body could've been specified in .dox files, but that would've
been more annoying than using @ingroup, which only applies to the
following entities in the current file (or something along those
lines). Additionally, this allows us to forgo duplicating the
group structure in headers, since the structure is now explicit
in .dox files. Yay!