string: removed null_tag{} ctor, corrected use.¹
1: The lack of a default ctor definition prohibited the use of
StringRef in FixedArray, which was promising to be useful. Empty
reference as the init state is the only reasonable init state
anyhow. This did highlight one issue with KVS: there is no
null_tag ctor for KVS, so constructing a KVS with {null_tag{}}
resulted in the string value ctor being used instead of the
(most likely intended) non-existent null value ctor. This KVS
ctor will not be added, as null value is, like empty StringRef,
the reasonable init state.