resource: took ResourceMetadata from tool_build.
resource/resource_metadata: renamed output_path() to compiled_path().

resource:
	added missing dox for resource_metadata.

serialization/resource:
	new support group;
	added serialization for ResourceMetadata.

tool_build:
	split ResourceMetadata into core metadata and ResourceCompilerMetadata;
	hacked up conversion for ResourceCompilerMetadata.¹

tool_build/package_compiler: changed manifest format (version 2).²

1: This should go away. Clang and GCC think ResourceCompilerMetadata
isn't standard layout when it derives ResourceMetadata, but from
my reading of the constraint on cppreference, it should be. The
structures are joined for convenience, and I'd like for that to be
the final case, but they may end up separate. It depends on how much
information is shared between ResourcePackage and PackageCompiler.

2: No more conditional serialization based on whether the metadata
is a "hole". Better to have predictable format for ResourcePackage,
and it's simpler code for PackageCompiler's manifest.