collection/types: constrain value types by TOGO_CONSTRAIN_POD_COLLECTION(T).¹
1: As noble as it is, there comes a time when you actually need
collections of non-POD types. This is currently uncommon, so the
solution here is to put the onus of lifetime management on the user.
It may be possible to do all lifetime management and permit any type
within collections (where management of POD types is optimized away
to NOPs), but that constitutes a considerable increase in complexity,
for little gain (so far).