lib/core/collection: moved struct defs into their interface namespaces; imported defs into root namespace.¹
1: To allow ADL on the collection interfaces. This is an experiment that might
be extended to all other interfaces. This allows unqualified incantation of
interface functions, but togo itself is unlikely to do so. Users can choose
to fully-qualify.

There are a few benefits I can see immediately:

  1. Don't have to specify interface name (obvious).
  2. Don't have to import root namespace (though most should).
  3. As a side-effect of (1): variable names can be more descriptive without
     causing super-mega-long incantations when using the interface.
  4. Free- and member-function interface can be defined in the same block.

(2) and (4) are minor. (1) and (3) can be significant, but not as significant
in this case, except when multiple incantations occur in a single line (which
certainly exist in this codebase).

Long names, such as those in lib/game/resource, would show the most benefit.

Some downsides of not fully-qualifying are:

  1. Can't find use of an interface by grepping (i.e., "<interface>::").
  2. Ambiguity forcing full qualification, causing convention exceptions.
  3. Accidentally using the wrong function.

(1) is unfortunate. Being able to use simple text processing to find calls is
really handy in several situations, most frequent of which are renaming and
debugging. (2) and (3) should be very minor. togo does not use TMP very much,
and any instance of it would not match before the definitions found through
ADL. What would be more annoying is the reverse: having to fully-qualify a
root-namespace call. In other words: (2) and (3) are unlikely.