lib/core/kvs/io/parser_new: new KVS parser using the parser interface.¹
1: Error reporting is completely non-functional, so this isn't
production ready (hence the old interface remaning, for now).
It's going to take some more work on the parser interface to allow
for true failure. It'll probably be a result handler parser sitting
above terminal parsers, where there is knowledge of what really
constitutes a failure... I am vaguely feeling that it's actually
already fine and that I'm just not paying attention to the result
flow here.